add useLogout hook
